Updated the Qty for Decimal status:Success

+++ app/src/main/java/com/example/mpos/data/deals/json/AddOnMenu.kt
@@ -7,6 +7,6 @@ data class AddOnMenu(
     @SerializedName("description") val description: String,
     @SerializedName("itemlist") val itemList: List<ItemList>,
     @SerializedName("MenuCode") val menuCode: String,
-    @SerializedName("Price") val price: Int,
+    @SerializedName("Price") val price: Double,
     @SerializedName("Type") val type: String
 )
