+++ app/src/main/java/com/example/mpos/data/occupied/json/OccupiedTableItem.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 
 data class OccupiedTableItem(
     @SerializedName("foodAmt") val foodAmt: Double,
-    @SerializedName("foodQty") val foodQty: Int,
+    @SerializedName("foodQty") val foodQty: Double,
     @SerializedName("itemMaster") val itemMaster: ItemMaster
 )
